refactor(ai-chat): reorganize module structure
Closes #93
