+++ app/ai-chat/src/app/menus.rs
