+++ app/ai-chat/src/components.rs
@@ -6,6 +6,7 @@ pub mod chat_form;
 pub mod delete_confirm;
 pub mod hotkey_input;
 pub mod message;
+pub(crate) mod search_list;
 pub mod template_edit_dialog;
 
 pub(crate) fn init(cx: &mut App) {
