+++ app/ai-chat/src/views/home/tabs/template_list.rs
@@ -1,6 +1,6 @@
 use crate::{
     assets::IconName,
-    components::template_edit_dialog::open_add_template_dialog,
+    components::{search_list, template_edit_dialog::open_add_template_dialog},
     database::{ConversationTemplate, Db},
     errors::AiChatResult,
     i18n::I18n,
@@ -8,7 +8,7 @@ use crate::{
 };
 use gpui::{prelude::FluentBuilder, *};
 use gpui_component::{
-    ActiveTheme, Icon, IndexPath, Selectable, Sizable,
+    ActiveTheme, Icon, Selectable, Sizable,
     button::{Button, ButtonVariants},
     h_flex,
     input::{Enter, Input, InputEvent, InputState, MoveDown, MoveUp},
@@ -250,13 +250,11 @@ impl TemplateListView {
         let list = cx.new(move |cx| {
             let mut state =
                 ListState::new(TemplateListDelegate::new(templates, on_confirm), window, cx);
-            let has_items = {
+            {
                 let delegate = state.delegate_mut();
                 delegate.apply_query(&query);
-                !delegate.filtered_items.is_empty()
-            };
-            state.set_selected_index(has_items.then_some(IndexPath::default()), window, cx);
-            state.scroll_to_item(IndexPath::default(), ScrollStrategy::Top, window, cx);
+            }
+            search_list::select_first_if_any(&mut state, window, cx);
             state
         });
         Ok(list)
@@ -311,13 +309,11 @@ impl TemplateListView {
         };
 
         templates.update(cx, |state, cx| {
-            let has_items = {
+            {
                 let delegate = state.delegate_mut();
                 delegate.apply_query(&query);
-                !delegate.filtered_items.is_empty()
-            };
-            state.set_selected_index(has_items.then_some(IndexPath::default()), window, cx);
-            state.scroll_to_item(IndexPath::default(), ScrollStrategy::Top, window, cx);
+            }
+            search_list::select_first_if_any(state, window, cx);
         });
         cx.notify();
     }
@@ -344,23 +340,7 @@ impl TemplateListView {
         };
 
         templates.update(cx, |state, cx| {
-            let count = state.delegate().filtered_items.len();
-            if count == 0 {
-                state.set_selected_index(None, window, cx);
-                return;
-            }
-
-            let current = state.selected_index().map(|ix| ix.row).unwrap_or(0);
-            let next = if delta < 0 {
-                if current == 0 { count - 1 } else { current - 1 }
-            } else if current + 1 >= count {
-                0
-            } else {
-                current + 1
-            };
-            let next_ix = IndexPath::default().row(next);
-            state.set_selected_index(Some(next_ix), window, cx);
-            state.scroll_to_item(next_ix, ScrollStrategy::Top, window, cx);
+            search_list::move_selected(state, delta, window, cx);
         });
         cx.notify();
     }
@@ -375,11 +355,7 @@ impl TemplateListView {
         };
 
         templates.update(cx, |state, cx| {
-            let selected = state.selected_index();
-            state
-                .delegate_mut()
-                .set_selected_index(selected, window, cx);
-            state.delegate_mut().confirm(enter.secondary, window, cx);
+            search_list::confirm_selected(state, enter.secondary, window, cx);
         });
         cx.stop_propagation();
     }
